fix(groupStudy): 그룹스터디 api들 그룹스터디 관련 / 스터디 히스토리 관련 두가지로 분리
