+++ src/main/java/com/witherview/configuration/StompHandler.java
@@ -1,6 +1,7 @@
 package com.witherview.configuration;
 
-import com.witherview.groupPractice.GroupStudyService;
+import com.witherview.groupPractice.GroupStudy.GroupStudyService;
+import com.witherview.groupPractice.history.StudyHistoryService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.messaging.Message;
@@ -16,6 +17,7 @@ import org.springframework.stereotype.Component;
 @Component
 public class StompHandler implements ChannelInterceptor {
     private final GroupStudyService groupStudyService;
+    private final StudyHistoryService studyHistoryService;
 
     // websocket을 통해 들어온 요청이 처리 되기전 실행된다.
     @Override
@@ -28,7 +30,7 @@ public class StompHandler implements ChannelInterceptor {
 
             Long id = Long.parseLong(room);
             try {
-                if(type.equals("feedback")) groupStudyService.findStudyHistory(id);
+                if(type.equals("feedback")) studyHistoryService.findStudyHistory(id);
                 else groupStudyService.findRoom(id);
             } catch (Exception e) {
                 throw new MessagingException(e.getMessage());
